requirements.txt
absl-py==0.8.1
asgiref==3.2.7
astor==0.8.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.4
cachetools==3.1.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
colorama==0.4.1
decorator==4.4.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.5
django-heroku==0.3.1
django-sslserver==0.22
dnspython==1.16.0
entrypoints==0.3
eventlet==0.25.1
Flask==1.0
Flask-SocketIO==4.2.1
gast==0.2.2
gevent==1.4.0
google-auth==1.7.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
greenlet==0.4.15
grpcio==1.25.0
h5py==2.10.0
idna==2.8
importlib-metadata==0.23
ipykernel==5.1.3
ipython==7.10.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==0.24
jedi==0.15.1
Jinja2==2.8
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
kaggle==1.5.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
MarkupSafe==0.23
mistune==0.8.4
monotonic==1.5
more-itertools==8.0.0
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.2
numpy==1.17.4
oauthlib==3.1.0
opt-einsum==3.1.0
pandocfilters==1.4.2
parso==0.5.1
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.11.0
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
Pygments==2.5.2
pyjwt==1.7.1
pyrsistent==0.15.6
python-dateutil==2.8.1
python-engineio==3.11.1
python-slugify==4.0.0
python-socketio==4.4.0
pytz==2019.3
pywin32==227
pywinpty==0.5.7
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
Send2Trash==1.5.0
six==1.13.0
social-auth-app-django
sqlparse==0.3.1
tensorboard==2.0.2
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
tornado==6.0.3
tqdm==4.39.0
traitlets==4.3.3
urllib3==1.24.3
virtualenv==16.7.5
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.3
whitenoise==5.0.1
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==0.6.0
